+++ src/index.js
@@ -9,6 +9,7 @@ const {
 
 const {BadgeFactory} = require('gh-badges');
 const es6Templates = require('es6-template-strings');
+const {loadNycConfig} = require('@istanbuljs/load-nyc-config');
 
 const writeFile = promisify(writeFileOriginal);
 const unlink = promisify(unlinkOriginal);
@@ -17,20 +18,69 @@ const unlink = promisify(unlinkOriginal);
  * @param {CoveradgeOptions} cfg
  * @returns {void}
  */
-async function coveradge ({
-  format = 'svg', color = 'orange', template = 'flat',
-  output = 'coverage-badge',
-  coveragePath = './coverage/coverage-summary.json',
-  // eslint-disable-next-line no-template-curly-in-string
-  textTemplate = 'Coverage ${pct}%'
-}) {
+async function coveradge (cfg) {
+  const {
+    format = 'svg',
+    passColor = 'green',
+    failColor = 'orange',
+    template = 'flat',
+    output = 'coverage-badge',
+    coveragePath = './coverage/coverage-summary.json',
+    // eslint-disable-next-line no-template-curly-in-string
+    textTemplate = 'Coverage ${failingConditionPct}%'
+  } = cfg;
+  const conditions = (cfg.conditions || '').split(',');
+  const nycConfig = await loadNycConfig();
+
   // eslint-disable-next-line global-require, import/no-dynamic-require
   const coverageSummary = require(join(process.cwd(), coveragePath));
 
-  const {pct} = coverageSummary.total.statements;
+  const possibleConditions = ['statements', 'branches', 'lines', 'functions'];
+  const possibleConditionThresholds = possibleConditions.reduce(
+    (o, condition) => {
+      // Priority to CLI condition threshold, then to nyc, then default to 100
+      o[condition] = cfg[condition + 'Threshold'] ||
+        nycConfig[condition] || 100;
+      return o;
+    },
+    {}
+  );
+
+  const failsThreshold = (condition) => {
+    const threshold = possibleConditionThresholds[condition];
+    return coverageSummary.total[condition] < threshold;
+  };
+
+  const conditionsToCheck = conditions.length
+    // User only wishes to check certain conditions
+    ? [
+      ...new Set(conditions.filter((condition) => {
+        return possibleConditions.includes(condition);
+      }))
+    ]
+    : possibleConditions;
+
+  const failingCondition = conditionsToCheck.find(
+    (thr) => failsThreshold(thr)
+  );
+  const color = failingCondition
+    ? failColor
+    : passColor;
 
   const text = es6Templates(
-    textTemplate, {pct}
+    textTemplate, {
+      failingCondition,
+      failingConditionPct: coverageSummary.total[failingCondition],
+      ...(
+        // If user specified conditions, only allow these in template;
+        //   otherwise, allow all
+        conditions.length ? conditions : possibleConditions
+      ).reduce((o, condition) => {
+        o[condition + 'Pct'] = coverageSummary.total[condition].pct;
+        o[condition + 'Threshold'] = possibleConditionThresholds[condition];
+        return o;
+      }, {})
+    }
   );
 
   // Only CLI handles image conversion
@@ -46,15 +96,15 @@ async function coveradge ({
   const badge = bf.create(formatInfo);
 
   // eslint-disable-next-line prefer-named-capture-group
-  output = output.replace(/\.(png|svg)$/u, '');
+  const outputBase = output.replace(/\.(png|svg)$/u, '');
 
-  const svgFilePath = `${output}.svg`;
+  const svgFilePath = `${outputBase}.svg`;
   await writeFile(svgFilePath, badge);
   console.log('Finished writing temporary SVG file...');
 
   if (format !== 'svg') {
     // For PNG, need conversion as built-in doesn't work for non-SVG images
-    const out = openSync(`${output}.${format}`, 'a');
+    const out = openSync(`${outputBase}.${format}`, 'a');
     spawn('./node_modules/.bin/badge', [
       text,
       `:${color}`,